requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
Pillow>=8.3.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
albumentations>=1.0.0
tqdm>=4.62.0
opencv-python>=4.5.0
fastapi>=0.68.0
fastapi[standard]>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.0 